torpedo/requirements.txt
djangorestframework
django-pipeline

# development
webtest
django-webtest
